+++ cli.ts
@@ -13,7 +13,7 @@ import {
 } from "./cli/mod.ts";
 import "https://deno.land/std@0.179.0/dotenv/load.ts";
 
-export const version = "v0.3.7";
+export const version = "v0.3.8";
 
 const command = new Command()
 	.name("arkiver")