+++ cli/init/mod.ts
@@ -30,7 +30,7 @@ manifest
 	.addEntity(Balance)
 	.chain("mainnet", { blockRange: 100n })
 	.contract(erc20)
-	.addSources({ "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2": 16986910n })
+	.addSources({ "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2": 16987911n })
 	.addEventHandlers({ "Transfer": transferHandler });
 
 export default manifest.build();`;
@@ -73,7 +73,7 @@ import { Balance } from "./entities.ts";
 
 export const transferHandler: EventHandlerFor<typeof erc20, "Transfer"> =
 	async (
-		{ event, client, store },
+		{ event, client, store, contract },
 	) => {
 		const { from, to, value } = event.args;
 
@@ -82,12 +82,7 @@ export const transferHandler: EventHandlerFor<typeof erc20, "Transfer"> =
 		// store.retrieve() will return the value if it exists in the store, otherwise it will run the function and store the result
 		const decimals = await store.retrieve(
 			\`\${address}:decimals\`,
-			async () =>
-				await client.readContract({
-					abi: erc20,
-					functionName: "decimals",
-					address,
-				}),
+			contract.read.decimals,
 		);
 
 		// reduce rpc calls in case you have multiple events in the same block