+++ examples/simple/manifest.ts
@@ -9,7 +9,7 @@ manifest
 	.addEntity(Balance)
 	.chain("mainnet", { blockRange: 100n })
 	.contract(erc20)
-	.addSources({ "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2": 16986910n })
+	.addSources({ "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2": 16987011n })
 	.addEventHandlers({ "Transfer": transferHandler });
 
 export default manifest.build();